refactor raffle in to container and component
