+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Provider } from 'react-redux'
 import store from './redux/store'
 import chooseWinner from './redux/actionCreators'
-import Raffle from './Raffle'
+import Raffle from './containers/RaffleContainer'
 
 const App = React.createClass({
   handleClick() {
